VFS, RO FAT32

+++ kernel/src/util/mem.h
@@ -2,8 +2,8 @@
 
 #include <util/types.h>
 
-void memset(void* d, u8 v, size_t size);
-void memcpy(void* d, const void* s, size_t size);
+void memset(void* d, u8 v, u64 size);
+void memcpy(void* d, const void* s, u64 size);
 i32 memcmp(const void* lhs, const void* rhs, size_t count);
 
 void memset_aligned(void* d, u64 v, u64 size);
